scanner/requirements.txt
pytrends==4.9.2
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.3
mysql-connector-python==8.3.0
schedule==1.2.1
ollama==0.4.7
feedparser==6.0.11
